Fix initialization problem in some environments
The code in clj-uuid.node is now evaluated lazily.

Background: the code from 108 to 136 relies that the network and
hardware of the underlying system behaves "well". I some cases it does
not. We observed this with containers. As long as v1 UUIDs are not
created via this library the problematic code is not called.